// src/domain/types/catalog/items/product-variant-item-details.ts
import { Image, MediaGallery, TString, UserCon } from '@glosuite/shared';
import { MiniProductVariantOutput } from 'src/domain/dto/items';
import { VoucherItemOutput } from 'src/domain/dto/orders';
import { Category } from 'src/domain/entities/structures';
import { ProductType } from 'src/domain/enums/items';
import { ShippingClass } from 'src/domain/enums/orders';
import { ProductQuantity } from 'src/domain/interfaces';
import { StockValueModel } from 'src/domain/interfaces/analytics';
import { ProductItemDetails } from 'src/domain/interfaces/items';
import { VariantAttributeValueModel } from './variant-attribute-value.model';

export type ProductVariantItemDetails = {
  id: string;
  magentoId?: number;
  magentoSKU?: string;
  reference: string;
  title: TString;
  name?: string;
  shortDescription?: TString;
  description?: TString;
  shippingClass: ShippingClass;
  sku: string;
  thumbnail: Image;
  gallery?: MediaGallery[];
  productType: ProductType;
  productId: string;
  attributeValues?: VariantAttributeValueModel<any>[];
  children?: MiniProductVariantOutput[];
  productItems?: ProductItemDetails[];
  categories: Category[];
  quantity: ProductQuantity;
  salePrice: number;
  specialPrice?: VoucherItemOutput;
  purchaseCost: number;
  stockValue?: StockValueModel;
  rentPrice?: number;
  canBeSold: boolean;
  canBeRented?: boolean;
  canBePackaged?: boolean;
  mustBePackaged?: boolean;
  createdBy?: UserCon;
  createdAt: Date;
  lastUpdate: Date;
};